Cleanup config names
