+++ crates/cubecl-runtime/src/config/compilation.rs
@@ -1,0 +1,9 @@
+use super::logger::{BinaryLogLevel, LoggerConfig};
+
+#[derive(Default, Clone, Debug, serde::Serialize, serde::Deserialize)]
+pub struct CompilationConfig {
+    #[serde(default)]
+    pub logger: LoggerConfig<CompilationLogLevel>,
+}
+
+pub type CompilationLogLevel = BinaryLogLevel;
