+++ crates/cubecl-runtime/src/tune/util.rs
@@ -1,6 +1,6 @@
 use core::sync::atomic::{AtomicI32, Ordering};
 
-use crate::config::CubeGlobalConfig;
+use crate::config::GlobalConfig;
 
 /// Autotune levels:
 ///
@@ -44,12 +44,12 @@ pub fn anchor(x: usize, max: Option<usize>, min: Option<usize>, base: Option<usi
 fn load_autotune_level() -> u32 {
     let autotune_level = AUTOTUNE_LEVEL.load(Ordering::Relaxed);
     if autotune_level == -1 {
-        let config = CubeGlobalConfig::get();
+        let config = GlobalConfig::get();
         let level = match config.autotune.level {
-            crate::config::AutotuneLevel::Minimal => 0,
-            crate::config::AutotuneLevel::Medium => 1,
-            crate::config::AutotuneLevel::More => 2,
-            crate::config::AutotuneLevel::Full => 3,
+            crate::config::autotune::AutotuneLevel::Minimal => 0,
+            crate::config::autotune::AutotuneLevel::Medium => 1,
+            crate::config::autotune::AutotuneLevel::More => 2,
+            crate::config::autotune::AutotuneLevel::Full => 3,
         };
         AUTOTUNE_LEVEL.store(level, Ordering::Relaxed);
         level as u32
